+++ Radiant/Radiant/StaticMeshManager.cpp
@@ -241,10 +241,10 @@ void StaticMeshManager::CreateStaticMesh(Entity entity, const char * filename, s
 		throw msg;
 	}
 
-	mesh->AddAttributeStream(Mesh::AttributeType::Position, pos.size(), (float*)&pos[0], indices.size(), &indices[0]);
-	mesh->AddAttributeStream(Mesh::AttributeType::TexCoord, uvs.size(), (float*)&uvs[0], indices.size(), &indices[0]);
+	mesh->AddAttributeStream(Mesh::AttributeType::Position, static_cast<unsigned int>(pos.size()), (float*)&pos[0], static_cast<unsigned int>(indices.size()), &indices[0]);
+	mesh->AddAttributeStream(Mesh::AttributeType::TexCoord, static_cast<unsigned int>(uvs.size()), (float*)&uvs[0], static_cast<unsigned int>(indices.size()), &indices[0]);
 
-	mesh->AddBatch(0, indices.size());
+	mesh->AddBatch(0, static_cast<unsigned int>(indices.size()));
 	
 	mesh->CalcNTB();
 	uint32_t vertexBufferIndex = 0;
